Add Pokemon search component
Added a new Pokemon search component with a search input field and a search icon. The component is styled with a transparent background and a border radius of 10px. The search icon is a new SVG file added to the assets folder. Also, updated the reset.scss file to set the font-weight of all elements to 400 and the input element to 100. Finally, updated the variables.scss file to correct a typo in the variable name.